requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
google-generativeai==0.3.2
spotipy==2.23.0
textblob==0.17.1
requests>=2.31.0
transformers>=4.30.0
torch>=2.0.0
tensorflow>=2.13.0  # Required by transformers 